+++ css/style.css
@@ -1035,3 +1035,7 @@ fieldset {
 .hidden {
   display: none;
 }
+
+.ad-form__error-text {
+  color:  #ff5635;
+}
