+++ js/main.js
@@ -1,7 +1,7 @@
 import{ createAds } from './data.js';
 import { renderCard } from './card.js';
 import {activateForm} from './form.js';
-import './validate-form.js';
+import {addValidatorsToForm}from './validate-form.js';
 
 //Количество выводимых объявлений
 const COUNT_ADS = 10;
@@ -11,3 +11,5 @@ const generatedAds = createAds(COUNT_ADS);
 renderCard(generatedAds[0]);
 
 activateForm(true);
+
+addValidatorsToForm();
